+++ tasks/build-package.sh
@@ -2,6 +2,6 @@
 
 set -e
 rm -fR package
-npm run package:sass
-npm run package:scripts
+npm run compile:sass
+npm run compile:scripts
 cp -R src/nationalarchives package.json package-lock.json README.md LICENCE govuk-prototype-kit.config.json package