+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Renamed "message" component to "warning"
 - Changed all `#000` colours to `#010101`
 - The newsletter subscribe button in the footer is no longer accented
 - Added `role="group"` to date input component `<fieldset>`